+++ tests/management/test_finalize_xblockskill_tags.py
@@ -3,16 +3,18 @@
 Tests for the django management command `finalize_xblockskill_tag`.
 """
 import logging
-import responses
+
+from pytest import mark
+from testfixtures import LogCapture
 
 from django.core.management import call_command
-from testfixtures import LogCapture
-from pytest import mark
-from test_utils import factories
-from test_utils.testcase import TaxonomyTestCase
-from test_utils.constants import USAGE_KEY
-from taxonomy.models import XBlockSkillData
+from django.test import override_settings
+
 from taxonomy.exceptions import InvalidCommandOptionsError
+from taxonomy.models import XBlockSkillData
+from test_utils import factories
+from test_utils.constants import USAGE_KEY
+from test_utils.testcase import TaxonomyTestCase
 
 
 @mark.django_db
@@ -26,7 +28,6 @@ class FinalizeSkillTagsCommandTests(TaxonomyTestCase):
         super().setUp()
         self.mock_access_token()
 
-    @responses.activate
     def test_finalize_xblockskill_tags_without_unverified_skills(self):
         """
         Test that command only shows starting and completed logs
@@ -35,7 +36,7 @@ class FinalizeSkillTagsCommandTests(TaxonomyTestCase):
         unverified_skills = XBlockSkillData.objects.filter(verified=False)
         self.assertEqual(len(unverified_skills), 0)
         with LogCapture(level=logging.INFO) as log_capture:
-            call_command(self.command, min_votes=2, ratio_threshold=0.5)
+            call_command(self.command, min_verified_votes=2, ratio_verified_threshold=0.5)
             self.assertEqual(len(log_capture.records), 2)
             messages = [record.msg for record in log_capture.records]
             self.assertEqual(
@@ -46,7 +47,6 @@ class FinalizeSkillTagsCommandTests(TaxonomyTestCase):
                 ]
             )
 
-    @responses.activate
     def test_finalize_xblockskill_tags_below_minimum_votes(self):
         """
         Test that command only shows starting and completed logs
@@ -66,7 +66,7 @@ class FinalizeSkillTagsCommandTests(TaxonomyTestCase):
         xblock_skill.ignored_count = 0
         xblock_skill.save()
         with LogCapture(level=logging.INFO) as log_capture:
-            call_command(self.command, min_votes=2, ratio_threshold=0.5)
+            call_command(self.command, min_verified_votes=2, ratio_verified_threshold=0.5)
             self.assertEqual(len(log_capture.records), 2)
             messages = [record.msg for record in log_capture.records]
             self.assertEqual(
@@ -77,7 +77,6 @@ class FinalizeSkillTagsCommandTests(TaxonomyTestCase):
                 ]
             )
 
-    @responses.activate
     def test_finalize_xblockskill_tags_below_ratio_threshold(self):
         """
         Test that command only shows starting and completed logs
@@ -100,7 +99,7 @@ class FinalizeSkillTagsCommandTests(TaxonomyTestCase):
         xblock_skill.ignored_count = 3
         xblock_skill.save()
         with LogCapture(level=logging.INFO) as log_capture:
-            call_command(self.command, min_votes=2, ratio_threshold=0.5)
+            call_command(self.command, min_verified_votes=2, ratio_verified_threshold=0.5)
             self.assertEqual(len(log_capture.records), 2)
             messages = [record.msg for record in log_capture.records]
             self.assertEqual(
@@ -111,11 +110,9 @@ class FinalizeSkillTagsCommandTests(TaxonomyTestCase):
                 ]
             )
 
-    @responses.activate
-    def test_finalize_xblockskill_tags(self):
+    def test_finalize_xblockskill_tags_for_verification(self):
         """
-        Test that command shows starting, verified and completed
-        logs
+        Test that finalize_xblockskill_tags verifies skills with correct votes.
         """
         xblock = factories.XBlockSkillsFactory(usage_key=USAGE_KEY)
         xblock_skill = factories.XBlockSkillDataFactory(xblock=xblock)
@@ -130,7 +127,7 @@ class FinalizeSkillTagsCommandTests(TaxonomyTestCase):
         xblock_skill.ignored_count = 1
         xblock_skill.save()
         with LogCapture(level=logging.INFO) as log_capture:
-            call_command(self.command, min_votes=2, ratio_threshold=0.5)
+            call_command(self.command, min_verified_votes=2, ratio_verified_threshold=0.5)
             self.assertEqual(len(log_capture.records), 3)
             messages = [record.msg for record in log_capture.records]
             self.assertEqual(
@@ -141,88 +138,31 @@ class FinalizeSkillTagsCommandTests(TaxonomyTestCase):
                     'Xblockskill tags verification task is completed'
                 ]
             )
-        updated_xblockskill = XBlockSkillData.objects.all()[0]  # there's only one
+        updated_xblockskill = XBlockSkillData.objects.first()  # there's only one
         self.assertTrue(updated_xblockskill.verified)
 
-    @responses.activate
-    def test_finalize_xblockskill_tags_without_min_votes(self):
+    @override_settings(SKILLS_VERIFICATION_THRESHOLD=None)
+    def test_finalize_xblockskill_tags_without_settings(self):
         """
-        Test that command raises InvalidCommandOption error if
-        minimum votes is not set in the settings file or sent in
-        as a parameter
+        Test that command raises InvalidCommandOptionsError if any setting and
+        argument is missing.
+        """
+        with self.assertRaises(InvalidCommandOptionsError):
+            call_command(self.command)
+
+    def test_finalize_xblockskill_tags_with_no_votes(self):
+        """
+        Test that command does nothing if no votes are present for given block
         """
         xblock = factories.XBlockSkillsFactory(usage_key=USAGE_KEY)
-        xblock_skill = factories.XBlockSkillDataFactory(xblock=xblock)
+        factories.XBlockSkillDataFactory(xblock=xblock)
 
         # ensure xblockskilldata object is created
         unverified_skills = XBlockSkillData.objects.filter(verified=False)
-        self.assertEqual(len(unverified_skills), 1)
-
-        xblock_skill.verified_count = 1
-        xblock_skill.ignored_count = 0
-        xblock_skill.save()
-
-        error_str = 'Either configure MIN_VOTES_FOR_SKILLS in settings or pass with arg --min-votes'
-        with LogCapture(level=logging.INFO) as log_capture:
-            with self.assertRaisesRegex(InvalidCommandOptionsError, error_str):
-                call_command(self.command, ratio_threshold=0.5)
-            self.assertEqual(len(log_capture.records), 1)
-            message = log_capture.records[0].msg
-            # only the staring message will be shown
-            self.assertEqual(
-                message,
-                'Starting xblockskill tags verification task',
-            )
-
-    @responses.activate
-    def test_finalize_xblockskill_tags_without_ratio_threshold(self):
-        """
-        Test that command raises InvalidCommandOption error if the
-        ratio threshold is not set in the settings file or sent in
-        as a parameter
-        """
-        xblock = factories.XBlockSkillsFactory(usage_key=USAGE_KEY)
-        xblock_skill = factories.XBlockSkillDataFactory(xblock=xblock)
-
-        # ensure xblockskilldata object is created
-        unverified_skills = XBlockSkillData.objects.filter(verified=False)
-        self.assertEqual(len(unverified_skills), 1)
-
-        xblock_skill.verified_count = 1
-        xblock_skill.ignored_count = 0
-        xblock_skill.save()
-
-        error_str = 'Either configure RATIO_THRESHOLD_FOR_SKILLS in settings or pass with arg --ratio-threshold'
-        with LogCapture(level=logging.INFO) as log_capture:
-            with self.assertRaisesRegex(InvalidCommandOptionsError, error_str):
-                call_command(self.command, min_votes=2)
-            self.assertEqual(len(log_capture.records), 1)
-            message = log_capture.records[0].msg
-            # only the staring message will be shown
-            self.assertEqual(
-                message,
-                'Starting xblockskill tags verification task',
-            )
-
-    @responses.activate
-    def test_finalize_xblockskill_tags_without_any_counts(self):
-        """
-        Test that command only shows starting and completed logs
-        if the skills have neither verified not ignored counts
-        """
-        xblock = factories.XBlockSkillsFactory(usage_key=USAGE_KEY)
-        xblock_skill = factories.XBlockSkillDataFactory(xblock=xblock)
-
-        # ensure xblockskilldata object is created
-        unverified_skills = XBlockSkillData.objects.filter(verified=False)
-        self.assertEqual(len(unverified_skills), 1)
-
-        xblock_skill.verified_count = 0
-        xblock_skill.ignored_count = 0
-        xblock_skill.save()
+        self.assertEqual(unverified_skills.count(), 1)
 
         with LogCapture(level=logging.INFO) as log_capture:
-            call_command(self.command, min_votes=2, ratio_threshold=0.5)
+            call_command(self.command)
             self.assertEqual(len(log_capture.records), 2)
             messages = [record.msg for record in log_capture.records]
             self.assertEqual(
@@ -232,3 +172,67 @@ class FinalizeSkillTagsCommandTests(TaxonomyTestCase):
                     'Xblockskill tags verification task is completed'
                 ]
             )
+        self.assertEqual(unverified_skills.count(), 1)
+
+    def test_finalize_xblockskill_tags_for_blacklisting(self):
+        """
+        Test that finalize_xblockskill_tags blacklists highly ignored skills.
+        """
+        xblock = factories.XBlockSkillsFactory(usage_key=USAGE_KEY)
+        xblock_skill = factories.XBlockSkillDataFactory(xblock=xblock)
+
+        # ensure xblockskilldata object is created
+        unverified_skills = XBlockSkillData.objects.filter(verified=False, is_blacklisted=False)
+        self.assertEqual(len(unverified_skills), 1)
+
+        # Set the verified count and ignored count so that the ratio is above
+        # the RATIO_THRESHOLD_FOR_SKILLS
+        xblock_skill.verified_count = 2
+        xblock_skill.ignored_count = 20
+        xblock_skill.save()
+        with LogCapture(level=logging.INFO) as log_capture:
+            call_command(self.command)
+            self.assertEqual(len(log_capture.records), 3)
+            messages = [record.msg for record in log_capture.records]
+            self.assertEqual(
+                messages,
+                [
+                    'Starting xblockskill tags verification task',
+                    '[%s] skill tag for the xblock [%s] has been blacklisted',
+                    'Xblockskill tags verification task is completed'
+                ]
+            )
+        updated_xblockskill = XBlockSkillData.objects.first()  # there's only one
+        self.assertFalse(updated_xblockskill.verified)
+        self.assertTrue(updated_xblockskill.is_blacklisted)
+
+    def test_argument_has_priority(self):
+        """
+        Test that command argument has priority over settings.
+        """
+        xblock = factories.XBlockSkillsFactory(usage_key=USAGE_KEY)
+        xblock_skill = factories.XBlockSkillDataFactory(xblock=xblock)
+
+        # ensure xblockskilldata object is created
+        unverified_skills = XBlockSkillData.objects.filter(verified=False)
+        self.assertEqual(len(unverified_skills), 1)
+
+        # Set the verified count and ignored count so that the ratio is above
+        # the settings.SKILLS_VERIFICATION_THRESHOLD and settings.SKILLS_VERIFICATION_RATIO_THRESHOLD
+        xblock_skill.verified_count = 3
+        xblock_skill.ignored_count = 1
+        xblock_skill.save()
+        with LogCapture(level=logging.INFO) as log_capture:
+            # setting thresholds higher than settings
+            call_command(self.command, min_verified_votes=8, ratio_verified_threshold=0.9)
+            self.assertEqual(len(log_capture.records), 2)
+            messages = [record.msg for record in log_capture.records]
+            self.assertEqual(
+                messages,
+                [
+                    'Starting xblockskill tags verification task',
+                    'Xblockskill tags verification task is completed'
+                ]
+            )
+        updated_xblockskill = XBlockSkillData.objects.first()  # there's only one
+        self.assertFalse(updated_xblockskill.verified)